+++ lib/yaks/collection_resource.rb
@@ -38,7 +38,7 @@ module Yaks
     # whatever it gets as a single resource with links and subresources,
     # we just push the collection down one level.
     #
-    # Once inside subresources the HAL serializer does check if a resource
+    # Once inside subresources the HAL format does check if a resource
     # is a collection, since there it does make a distinction, and because
     # in that case it will iterate with each/map rather than calling subresources,
     # this doesn't cause infinite recursion. Not very pretty, needs looking at.