+++ lib/yaks/config.rb
@@ -7,7 +7,9 @@ module Yaks
       @default_format = :hal
       @policy_options = {}
       @primitivize    = Primitivize.create
-      @steps          = [ @primitivize ]
+      @steps          = [
+        @primitivize
+      ]
       DSL.new(self, &blk)
     end
 
@@ -15,13 +17,13 @@ module Yaks
       @policy_class.new(@policy_options)
     end
 
-    def serializer_class(opts, env)
+    def format_class(opts, env)
       if env.key? 'HTTP_ACCEPT'
         accept = Rack::Accept::Charset.new(env['HTTP_ACCEPT'])
-        mime_type = accept.best_of(Serializer.mime_types.values)
-        return Serializer.by_mime_type(mime_type) if mime_type
+        mime_type = accept.best_of(Format.mime_types.values)
+        return Format.by_mime_type(mime_type) if mime_type
       end
-      Serializer.by_name(opts.fetch(:format) { @default_format })
+      Format.by_name(opts.fetch(:format) { @default_format })
     end
 
     def format_name(opts)
@@ -45,9 +47,9 @@ module Yaks
       }
 
       mapper     = opts.fetch(:mapper) { policy.derive_mapper_from_object(object) }.new(context)
-      serializer = serializer_class(opts, env).new(format_options[format_name(opts)])
+      format = format_class(opts, env).new(format_options[format_name(opts)])
 
-      [ mapper, serializer, *steps ].inject(object) {|memo, step| step.call(memo) }
+      [ mapper, format, *steps ].inject(object) {|memo, step| step.call(memo) }
     end
     alias serialize call
   end