+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ### Development
 [full changelog](http://github.com/plexus/yaks/compare/v0.4.2...master)
 
+* Yaks::Serializer has been renamed to Yaks::Format
+
 * Yaks::Mapper#{map_attributes,map_links,map_subresource} signature has changed, they now are responsible for adding themselves to a resource instance.
 
 ```ruby